Updated sequence detection (V2)
Changed to only use the previous two instances of the pattern to try and stop it spamming water when it shouldn't. Also made sure it matched patterns of the same size rather than matching to part of another sequence (also removed the cropping of large sequences). Now beats Amish V1 and occasionally Erin V1 and Amish V3